requirements.txt
appdirs==1.4.4
audioread==2.1.9
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1607575130224/work
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.10
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz==0.11.0
dask==2021.10.0
decorator==5.1.1
fonttools==4.25.0
fsspec @ file:///opt/conda/conda-bld/fsspec_1642510437511/work
idna==3.3
idx2numpy==1.2.3
imagecodecs @ file:///tmp/build/80754af9/imagecodecs_1635529108216/work
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
joblib==1.1.0
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612281846643/work
librosa==0.8.1
llvmlite==0.38.0
locket==0.2.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1638289681807/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
munkres==1.1.4
networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
numba==0.55.0
numexpr @ file:///tmp/build/80754af9/numexpr_1640689833592/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095651905/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
partd @ file:///tmp/build/80754af9/partd_1618000087440/work
Pillow==8.4.0
pooch==1.5.2
pycparser==2.21
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1607645421828/work
PyYAML==6.0
requests==2.27.1
resampy==0.2.2
scikit-image==0.18.3
scikit-learn==1.0.2
scipy @ file:///tmp/build/80754af9/scipy_1641555004408/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
SoundFile==0.10.3.post1
threadpoolctl==3.0.0
tifffile @ file:///tmp/build/80754af9/tifffile_1627275862826/work
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
torch==1.10.1
torchaudio==0.10.1
torchinfo==1.6.3
torchvision==0.11.2
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.8
